CC-1: seperated logic and useCases
