+++ README.md
@@ -1,8 +1,9 @@
 CC-1 TODO:
   - More clearly define controllers, app logic and business logic
-    - Hooks as Controllers
-    - App logic for any JS that changes the UI or changes data in DB
-    - Business logic for any JS that generates results or data
-  - Reduce the ammount of functions im sure this can be simplified
+    - Hooks are not Controllers
+    - use case:Defines how the outside world interacts with the business logic
+    - logic: Core rules and policies that never change regardless of UI
+    - DB is part of the outer most ring
+  - Reduce the amount of functions im sure this can be simplified
   - Fix issues
-    - 
+    -
