+++ src/useCases/powerList/UpdateTask.ts
@@ -1,12 +1,12 @@
 import { Task, TaskList } from '@/types/powerList';
+import { updateTaskListStatus } from '@/logic/powerList'
 
 type UpdateTaskArgs = {
   currentTaskList: TaskList | null;
   setCurrentTaskList: (taskList: TaskList) => void;
-  updateTaskListStatus: (taskList: TaskList) => TaskList;
 };
 
-const UpdateTask = ({ currentTaskList, setCurrentTaskList, updateTaskListStatus }: UpdateTaskArgs) =>
+const UpdateTask = ({ currentTaskList, setCurrentTaskList }: UpdateTaskArgs) =>
   (taskId: string, text: string ) => {
     if (!currentTaskList) return;
